* Update view sort

* Add wrapper for IdentifyAndStoreActiveClonotypes

* Error checking for CommonFilters

* Update defaults

* Set baseUrl in script

* Do not storeStimLevelData for TRA

* Do not force seurat object name to match readset

* Do not force seurat object name to match readset

* Drop CS-Core and Tricycle

* Expand data cleanup

* Minor code cleanup

* Add subadapter + reflection pattern to ExtendedVariantAdapter in order to delete VCFTabixAdapter (#353)

* Add subadapter + reflection pattern to ExtendedVariantAdapter in order to delete VCFTabixAdapter

* Add API wrapper for getMetadata and getHeader

---------


* Correct HTML syntax

* Expand study triggers and update cohort fields

* Expose getter for StudiesTriggerFactory

* Clean up trigger/customizer layer code

* Create fields to coalesce name/label for studies

* Add null check

* Switch default

* Bugfix to StudiesTriggerFactory

* Update default

* Improve SnpEff index check

* Switch sequence init tasks to use webserver-high-priority

* Build short delay into github triggers to aid cross-repo commits

* Switch ETLs to log row count discrepancies

* Update dependencies

* Support sawfish --sample-csv arg

* Option to create readsets from SRA (#355)

* Option to create readsets from SRA
* Test fix

* Error check

* Add nimble/bulk step

* Allow nimble step to use cached barcodes

* Bugfix to NimbleAlignmentStep

* Bugfix to NimbleAlignmentStep

* Bugfix to NimbleAlignmentStep

* Bugfix to NimbleAlignmentStep

* Add CD4_Activation_Axis

* Better support readsets created directly from SRA

* Expand BAM header

* Bugfix to RestoreSraDataHandler

* Bugfix to RestoreSraDataHandler for new SRA datasets

* Bugfix to RestoreSraDataHandler for new SRA datasets

* Bugfix to RestoreSraDataHandler for new SRA datasets

* Bugfix to RestoreSraDataHandler for new SRA datasets

* Reduce logging

* Update sawfish install

* Better error handling

* Updates to Save10xBarcodes

* Expand StudyMetadata cohorts

* Throw exception when existing file present

* Improve resume for ReadsetInitTask

* Improve resume for ReadsetInitTask

* Bugfix to Save10xBarcodes

* Bugfix to handling of 10x barcodes

* Switch nimble/CR barcodes to CB alone

* Omit writing to 10x barcodes

* Add another coalesce() term in case name and label are blank

* Bugfix to study import

* Fix typo

* Add manage freezer button and expand TagPcrSummaryStep

* Change the pattern of nimble fastq-to-bam

* Add check for integers

* Expand ETL/presentation of SIV data

* Expand ETL/presentation of SIV data

* Int -> long

* npm dependency updates

* Build fix

---------

+++ SequenceAnalysis/src/org/labkey/sequenceanalysis/pipeline/ReadsetInitTask.java
@@ -430,6 +430,10 @@ public class ReadsetInitTask extends WorkDirectoryTask<ReadsetInitTask.Factory>
                         moveInputToAnalysisDir(compressed, job, actions, unalteredInputs, outputFiles);
                     }
                 }
+                else
+                {
+                    job.getLogger().debug("Input file does not exist, may have already been moved: " + input.getPath());
+                }
             }
         }
         else
@@ -450,23 +454,40 @@ public class ReadsetInitTask extends WorkDirectoryTask<ReadsetInitTask.Factory>
             File outputDir = job.getAnalysisDirectory();
             File output = new File(outputDir, input.getName());
             job.getLogger().debug("Destination: " + output.getPath());
+            boolean alreadyMoved = false;
             if (output.exists())
             {
+                job.getLogger().debug("output already exists");
                 if (unalteredInputs != null && unalteredInputs.contains(output))
                 {
                     job.getLogger().debug("\tThis input was unaltered during normalization and a copy already exists in the analysis folder so the original will be discarded");
                     input.delete();
-                    TaskFileManagerImpl.swapFilesInRecordedActions(job.getLogger(), input, output, actions, job, null);
-                    return;
+                    alreadyMoved = true;
                 }
                 else
                 {
-                    output = new File(outputDir, FileUtil.getBaseName(input.getName()) + ".orig.gz");
-                    job.getLogger().debug("\tA file with the expected output name already exists, so the original will be renamed: " + output.getPath());
+                    if (input.length() == output.length() && input.lastModified() == output.lastModified())
+                    {
+                        job.getLogger().info("Output exists, but has the same size/modified timestamp. Deleting original");
+                        input.delete();
+                        alreadyMoved = true;
+                    }
+                    else if (input.exists() && input.length() > output.length() && input.lastModified() == output.lastModified())
+                    {
+                        job.getLogger().info("Output exists with same timestamp, but with smaller file size. This probably indicates a truncated/failed copy. Deleting this file.");
+                        output.delete();
+                    }
+                    else
+                    {
+                        throw new PipelineJobException("A file with the expected output name already exists: " + output.getPath());
+                    }
                 }
             }
 
-            FileUtils.moveFile(input, output);
+            if (!alreadyMoved)
+            {
+                FileUtils.moveFile(input, output);
+            }
             if (!output.exists())
             {
                 throw new PipelineJobException("Unable to move file: " + input.getPath());
@@ -488,7 +509,7 @@ public class ReadsetInitTask extends WorkDirectoryTask<ReadsetInitTask.Factory>
 
             TaskFileManagerImpl.swapFilesInRecordedActions(job.getLogger(), input, output, actions, job, null);
         }
-        catch (IOException e)
+        catch (Exception e)
         {
             throw new PipelineJobException(e);
         }
